server/mcp_server_tls/pyproject.toml
[project]
name = "mcp-server-tls"
version = "0.1.0"
description = "TLS MCP server project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.4.1",
    "python-dotenv>=1.0.1",
    "volcengine>=1.0.175",
]

[project.scripts]
mcp-server-tls = "mcp_server_tls.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_tls"]